requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.6
django-cors-headers==3.7.0
django-extensions==3.1.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
flake8==3.8.4
flake8-isort==4.0.0
isort==5.7.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2021.1
regex==2020.11.13
sqlparse==0.4.1
testfixtures==6.17.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
